+++ src/main/java/backend/zip/service/event/EventServiceImpl.java
@@ -1,0 +1,4 @@
+package backend.zip.service.event;
+
+public interface EventServiceImpl {
+}
